+++ mopidy/internal/storage.py
