+++ mopidy/local/json.py
@@ -9,6 +9,7 @@ import sys
 import mopidy
 
 from mopidy import compat, local, models
+from mopidy import internal
 from mopidy.internal import timer
 from mopidy.local import search, storage, translator
 
@@ -98,7 +99,7 @@ class JsonLibrary(local.Library):
                     self._json_file)
                 self._tracks = {}
             else:
-                library = models.storage.load(self._json_file)
+                library = internal.storage.load(self._json_file)
                 self._tracks = dict((t.uri, t) for t in
                                     library.get('tracks', []))
         with timer.time_logger('Building browse cache'):
@@ -166,9 +167,9 @@ class JsonLibrary(local.Library):
         self._tracks.pop(uri, None)
 
     def close(self):
-        models.storage.save(self._json_file,
-                            {'version': mopidy.__version__,
-                             'tracks': self._tracks.values()})
+        internal.storage.save(self._json_file,
+                              {'version': mopidy.__version__,
+                               'tracks': self._tracks.values()})
 
     def clear(self):
         try:
