+++ mopidy/models/__init__.py
@@ -1,7 +1,6 @@
 from __future__ import absolute_import, unicode_literals
 
 from mopidy import compat
-from mopidy.internal import validation
 from mopidy.models import fields
 from mopidy.models.immutable import ImmutableObject, ValidatedImmutableObject
 from mopidy.models.serialize import ModelJSONEncoder, model_json_decoder
@@ -9,8 +8,7 @@ from mopidy.models.serialize import ModelJSONEncoder, model_json_decoder
 __all__ = [
     'ImmutableObject', 'Ref', 'Image', 'Artist', 'Album', 'track', 'TlTrack',
     'Playlist', 'SearchResult', 'model_json_decoder', 'ModelJSONEncoder',
-    'ValidatedImmutableObject', 'HistoryTrack', 'HistoryState', 'MixerState',
-    'PlaybackState', 'TracklistState']
+    'ValidatedImmutableObject']
 
 
 class Ref(ValidatedImmutableObject):
@@ -366,108 +364,3 @@ class SearchResult(ValidatedImmutableObject):
 
     # The albums matching the search query. Read-only.
     albums = fields.Collection(type=Album, container=tuple)
-
-
-class HistoryTrack(ValidatedImmutableObject):
-    """
-    A history track. Wraps a :class:`Ref` and it's timestamp.
-
-    :param timestamp: the timestamp
-    :type timestamp: int
-    :param track: the track reference
-    :type track: :class:`Ref`
-    """
-
-    # The timestamp. Read-only.
-    timestamp = fields.Integer()
-
-    # The track reference. Read-only.
-    track = fields.Field(type=Ref)
-
-
-class HistoryState(ValidatedImmutableObject):
-    """
-    State of the history controller.
-    Internally used for import/export of current state.
-
-    :param history: the track history
-    :type history: list of :class:`HistoryTrack`
-    """
-
-    # The tracks. Read-only.
-    history = fields.Collection(type=HistoryTrack, container=tuple)
-
-
-class MixerState(ValidatedImmutableObject):
-    """
-    State of the mixer controller.
-    Internally used for import/export of current state.
-
-    :param volume: the volume
-    :type volume: int
-    """
-
-    # The volume. Read-only.
-    volume = fields.Integer(min=0, max=100)
-
-
-class PlaybackState(ValidatedImmutableObject):
-    """
-    State of the playback controller.
-    Internally used for import/export of current state.
-
-    :param tlid: current track tlid
-    :type tlid: int
-    :param position: play position
-    :type position: int
-    :param state: playback state
-    :type state: :class:`validation.PLAYBACK_STATES`
-    """
-
-    # The tlid of current playing track. Read-only.
-    tlid = fields.Integer(min=1)
-
-    # The playback position. Read-only.
-    position = fields.Integer(min=0)
-
-    # The playback state. Read-only.
-    state = fields.Field(choices=validation.PLAYBACK_STATES)
-
-
-class TracklistState(ValidatedImmutableObject):
-
-    """
-    State of the tracklist controller.
-    Internally used for import/export of current state.
-
-    :param repeat: the repeat mode
-    :type repeat: bool
-    :param consume: the consume mode
-    :type consume: bool
-    :param random: the random mode
-    :type random: bool
-    :param single: the single mode
-    :type single: bool
-    :param next_tlid: the single mode
-    :type next_tlid: bool
-    :param tl_tracks: the single mode
-    :type tl_tracks: list of :class:`TlTrack`
-    """
-
-    # The repeat mode. Read-only.
-    repeat = fields.Boolean()
-
-    # The consume mode. Read-only.
-    consume = fields.Boolean()
-
-    # The random mode. Read-only.
-    random = fields.Boolean()
-
-    # The single mode. Read-only.
-    single = fields.Boolean()
-
-    # The repeat mode. Read-only.
-    next_tlid = fields.Integer(min=0)
-
-    # The list of tracks. Read-only.
-    tl_tracks = fields.Collection(type=TlTrack, container=tuple)
