+++ mopidy/models/serialize.py
@@ -4,9 +4,7 @@ import json
 
 from mopidy.models import immutable
 
-_MODELS = ['Ref', 'Artist', 'Album', 'Track', 'TlTrack', 'Playlist',
-           'HistoryTrack', 'HistoryState', 'MixerState', 'PlaybackState',
-           'TracklistState']
+_MODELS = ['Ref', 'Artist', 'Album', 'Track', 'TlTrack', 'Playlist']
 
 
 class ModelJSONEncoder(json.JSONEncoder):
@@ -46,4 +44,7 @@ def model_json_decoder(dct):
         model_name = dct.pop('__model__')
         if model_name in _MODELS:
             return getattr(models, model_name)(**dct)
+        from mopidy import internal
+        if model_name in internal.models._MODELS:
+            return getattr(internal.models, model_name)(**dct)
     return dct
