+++ tests/core/test_mixer.py
@@ -7,7 +7,7 @@ import mock
 import pykka
 
 from mopidy import core, mixer
-from mopidy.models import MixerState
+from mopidy.internal.models import MixerState
 from tests import dummy_mixer
